Объединить тесты
